Extract FakeTitleElementService
