+++ src/Feature/AbstractManager.ts
@@ -6,10 +6,7 @@ import { Feature } from "@src/enum";
 @injectable()
 export default abstract class AbstractManager extends AbstractFeature<Feature> implements ManagerFeatureInterface {
     disable(): void {
-        if (this.isEnabled()) {
-            this.removeFakeTitleElements();
-            this.doDisable();
-        }
+        this.isEnabled() && this.doDisable();
     }
 
     enable(): void {
@@ -38,107 +35,4 @@ export default abstract class AbstractManager extends AbstractFeature<Feature> i
     protected abstract doUpdate(path: string): Promise<boolean>;
 
     protected abstract doRefresh(): Promise<{ [k: string]: boolean }>;
-
-    private fakeTitleElementMap = new Map<HTMLElement, HTMLElement>();
-
-    protected addFakeTitleElement(originalElement: HTMLElement, title: string): void {
-        if (!originalElement) {
-            return;
-        }
-
-        const container = originalElement.parentElement;
-
-        if (!container || !title) {
-            this.removeFakeTitleElement(originalElement);
-            return;
-        }
-
-        let fakeTitleElement = this.fakeTitleElementMap.get(originalElement);
-
-        if (fakeTitleElement?.innerText === title) {
-            return;
-        }
-
-        if (!fakeTitleElement) {
-            fakeTitleElement = document.createElement("div");
-            this.fakeTitleElementMap.set(originalElement, fakeTitleElement);
-            fakeTitleElement.className = originalElement.className;
-            fakeTitleElement.innerText = title;
-
-            originalElement.addEventListener("mouseenter", this.mouseEnterHandler);
-            originalElement.addEventListener("mouseleave", this.mouseLeaveHandler);
-            originalElement.addEventListener("blur", this.blurHandler);
-
-            fakeTitleElement.addEventListener("click", () => {
-                fakeTitleElement.hidden = true;
-                originalElement.hidden = false;
-                originalElement.focus();
-            });
-            container.insertBefore(fakeTitleElement, originalElement);
-        }
-
-        fakeTitleElement.innerText = title;
-        this.showFakeTitleElement(originalElement);
-    }
-
-    private showFakeTitleElement(originalElement: HTMLElement) {
-        const fakeTitleElement = this.fakeTitleElementMap.get(originalElement);
-        originalElement.hidden = true;
-        fakeTitleElement.hidden = false;
-    };
-
-    private mouseEnterHandler = ((event: MouseEvent) => {
-        const srcElement = event.target as HTMLElement
-        srcElement.dataset['hovered'] = 'true';
-    }).bind(this);
-
-    private mouseLeaveHandler = ((event: MouseEvent) => {
-        const srcElement = event.target as HTMLElement
-        srcElement.dataset['hovered'] = 'false';
-        setTimeout(() => {
-            this.restoreFakeTitleElement(srcElement);
-        }, 500);
-    }).bind(this);
-
-    private restoreFakeTitleElement(srcElement: HTMLElement) {
-        if (srcElement.dataset['hovered'] === 'false') {
-            if (document.activeElement !== srcElement) {
-                this.showFakeTitleElement(srcElement);
-            } else {
-                setTimeout(() => {
-                    this.restoreFakeTitleElement(srcElement);
-                }, 500);
-            }
-        }
-    }
-
-    private blurHandler = ((event: MouseEvent) => {
-        const srcElement = event.target as HTMLElement
-        this.showFakeTitleElement(srcElement);
-    }).bind(this);
-
-    private removeFakeTitleElements(): void {
-        for (const originalElement of this.fakeTitleElementMap.keys()) {
-            this.removeFakeTitleElement(originalElement);
-        }
-    }
-
-    private removeFakeTitleElement(originalElement: HTMLElement): void {
-        if (!originalElement) {
-            return;
-        }
-
-        originalElement.hidden = false;
-
-        const fakeTitleElement = this.fakeTitleElementMap.get(originalElement);
-
-        if (fakeTitleElement) {
-            fakeTitleElement.remove();
-            this.fakeTitleElementMap.delete(originalElement);
-        }
-
-        originalElement.removeEventListener("mouseenter", this.mouseEnterHandler);
-        originalElement.removeEventListener("mouseleave", this.mouseLeaveHandler);
-        originalElement.removeEventListener("blur", this.blurHandler);
-    }
 }
